+++ modules/terminate-agent-hook/iam.tf
@@ -40,7 +40,7 @@ data "aws_iam_policy_document" "lambda" {
       "ec2:DescribeInstances",
       "ec2:DescribeTags",
       "ec2:DescribeRegions",
-      "ec2:DescribeInstanceStatus"
+      "ec2:DescribeInstanceStatus",
     ]
     resources = ["*"]
     effect    = "Allow"
@@ -71,6 +71,7 @@ data "aws_iam_policy_document" "lambda" {
     ]
     resources = [var.asg_arn]
   }
+
   statement {
     sid = "GitLabRunnerLifecycleTerminateLogs"
     actions = [
@@ -85,16 +86,70 @@ data "aws_iam_policy_document" "lambda" {
       "${aws_cloudwatch_log_group.lambda.arn}:log-stream:*"
     ]
   }
+
+  statement {
+    sid = "SSHKeyHousekeepingList"
+
+    effect = "Allow"
+    actions = [
+      "ec2:DescribeKeyPairs"
+    ]
+    resources = ["*"]
+  }
+
+  # separate statement due to the condition
+  statement {
+    sid = "SSHKeyHousekeepingDelete"
+
+    effect = "Allow"
+    actions = [
+      "ec2:DeleteKeyPair"
+    ]
+    resources = ["*"]
+    condition {
+      test     = "StringLike"
+      variable = "ec2:KeyPairName"
+      values   = ["runner-*"]
+    }
+  }
+}
+
+data "aws_iam_policy_document" "spot_request_housekeeping" {
+  statement {
+    sid = "SpotRequestHousekeepingList"
+
+    effect = "Allow"
+    actions = [
+      "ec2:CancelSpotInstanceRequests",
+      "ec2:DescribeSpotInstanceRequests"
+    ]
+    # I didn't found any condition to limit the access
+    resources = ["*"]
+  }
 }
 
 resource "aws_iam_policy" "lambda" {
-  name   = "${var.name_iam_objects}-${var.name}"
+  name   = "${var.name_iam_objects}-${var.name}-lambda"
   path   = "/"
   policy = data.aws_iam_policy_document.lambda.json
-  tags   = var.tags
+
+  tags = var.tags
 }
 
 resource "aws_iam_role_policy_attachment" "lambda" {
   role       = aws_iam_role.lambda.name
   policy_arn = aws_iam_policy.lambda.arn
+}
+
+resource "aws_iam_policy" "spot_request_housekeeping" {
+  name   = "${var.name_iam_objects}-${var.name}-cancel-spot"
+  path   = "/"
+  policy = data.aws_iam_policy_document.spot_request_housekeeping.json
+
+  tags = var.tags
+}
+
+resource "aws_iam_role_policy_attachment" "spot_request_housekeeping" {
+  role       = aws_iam_role.lambda.name
+  policy_arn = aws_iam_policy.spot_request_housekeeping.arn
 }