+++ modules/terminate-agent-hook/locals.tf
@@ -1,0 +1,3 @@
+locals {
+  lambda_timeout = 30
+}